Fix Keystone API version detection
Keystone has supported v3 API for a long time and the v3
endpoint is available regardless of which API version the
the deployment is configured to use.

Rework the auto-detection and base it on some known facts
about Keystone versions and by reading charm configuration.

Add deprecation warning for referencing rc file by `novarc`
or `novarc_auto` name, preparing for rename to `openrc`.

Fixes LP: #1760806