WIP - Add Pipfile as a mechanism to install env for networking scripts
WORK IN PROGRESS - DO NOT MERGE YET

The networking scripts neutron-ext-net-ksv3 and neutron-tenant-net-ksv3
are python scripts that require python-neutronclient and
python-keystoneclient to be installed.  If they are not, then the
scripts will fail leading to a poor user experience.  This patchset
provides an option to install a pipenv environment where the commands
can be successfully run.

Related-Bug: #1671929