Removed rabbitmq:glance-simplestreams-sync relation (#188)
Remove rabbitmq:glance-simplestreams-sync relation

This relation is not supported anymore in OpenStack 20.08
https://docs.openstack.org/charm-guide/latest/2008.html

Also add a new Telegraf-Prometheus-Grafana overlay.
This overlay adds a 4th machine to the openstack-base overlay 
containing Prometheus and Grafana. The constraints will need
to be edited per use case. Currently there is an issue with
prometheus causing it to not show newly added targets without
reloading the service.
https://bugs.launchpad.net/charm-prometheus2/+bug/1891942
The service can be reloaded by `sudo snap restart prometheus`
on the prometheus host. At the moment the
prometheus-openstack-exporter is also not working.
https://bugs.launchpad.net/charm-prometheus-openstack-exporter/+bug/1823011

Co-authored-by: Aurelien Lourot <aurelien.lourot@canonical.com>