Drop openstack-dpdk bundle.
The dpdk setup is tested by zaza[0] and this bundle is Xenial based
which is End-Of-Life, the charms already dropped support for Xenial, so
the bundle is not usable in its current and provides no value to try to
fix it (e.g. pinning to old versions of the charms).

[0] https://github.com/openstack-charmers/zaza-openstack-tests/pull/746