Drop legacy charm urls from bundles
Previously, charm urls could reference Ubuntu series as part of the
charm that was being delivered (e.g. <team>/<series>/<charm>). Some
of the legacy bundles in this repository used this particular URL
schema to target specific series of charms.

However, this was deprecated as the series was added to the charm
metadata and the new charmhub store does not support this particular
API and subsequently causes the lint tests to fail as a result.

This removes all of these urls from the bundles in order to allow
the lint checks to pass.

Signed-off-by: Billy Olsen <billy.olsen@gmail.com>