enable-ml2-port-security for Yoga+
Neutron's port security extension allows operators to turn on and off
the port security on per port or per network basis. It brings
flexibility to network port management and it's expected to be available
where the extension is available, OVS/OVN deployments in mind. Let's
turn it on for new deployments.

Also, it's must-have for Octavia as per the charm deployment guide so
make sure it's enabled in the Octavia overlay bundle.